requirements.txt
# Requirements for BrainPy Wendling Model
# Environment: brainpy311 conda environment (Python 3.11)

# === Core Dependencies ===
# JAX and BrainPy (GPU support - CUDA 12 recommended)
# Note: For non-CUDA systems, you may need to install jax[cpu] instead.
jax==0.9.0
jaxlib==0.9.0
jax-cuda12-pjrt==0.9.0
jax-cuda12-plugin==0.9.0
brainpy[jax]==2.7.6

# Scientific computing
numpy==2.3.5
scipy==1.17.0
matplotlib==3.10.8

# === Optimization ===
nevergrad==1.0.12
tqdm==4.67.1

# === Data Handling ===
pandas==2.3.3
h5py==3.15.1

# === Configuration & Utilities ===
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# === Development & Testing ===
pytest>=7.0.0
ruff>=0.1.0
mkdocs>=1.5.0

# === Package Installation ===
-e .